move token voting governance components to own directory
