+++ src/components/DaoCreator/tokenVotingGovernance/SubsidiaryFunding/types/index.ts
@@ -1,7 +1,7 @@
 import {
   TreasuryAssetFungible,
   TreasuryAssetNonFungible,
-} from '../../../../providers/treasury/types';
+} from '../../../../../providers/treasury/types';
 
 export type TokenToFund = {
   asset: TreasuryAssetFungible;
